+++ .gitignore
@@ -71,3 +71,5 @@ typings/
 
 # Serverless directories
 .serverless
+
+.vscode
