+++ config/webpack/webpack.front.config.js
@@ -1,9 +1,9 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: path.resolve(__dirname, '../../src/front/index.js'),
   output: {
-    path: path.resolve(__dirname, 'dist'),
+    path: path.resolve(__dirname, '../../dist'),
     publicPath: 'dist',
     filename: 'reddit.bundle.js',
   },
