+++ package.json
@@ -4,8 +4,8 @@
   "description": "A Reddit clone based on React.",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --mode development",
-    "build": "webpack --progress"
+    "start": "webpack-dev-server --config config/webpack/webpack.front.config.js --mode development",
+    "build": "webpack --config config/webpack/webpack.front.config.js --progress"
   },
   "license": "MIT",
   "dependencies": {
