+++ src/front/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App/App';
+import App from './app/components/app/App';
 
 require('./styles/styles.scss');
