+++ src/front/styles/mixins/_bootstrap-text-hide.scss
