+++ src/front/styles/mixins/_bootstrap-text-truncate.scss
