+++ src/front/styles/styles.scss
